requirements.txt
flask
flask-cors
sentence-transformers
torch
python-dotenv
